+++ src/main/java/prgrms/neoike/common/config/JpaAuditingConfig.java
@@ -1,7 +1,9 @@
 package prgrms.neoike.common.config;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
+@EnableJpaAuditing
 public class JpaAuditingConfig {
 }
