+++ src/main/java/prgrms/neoike/controller/mapper/DrawMapper.java
@@ -5,8 +5,8 @@ import prgrms.neoike.controller.dto.drawdto.DrawSaveRequest;
 import prgrms.neoike.service.dto.drawdto.ServiceDrawSaveDto;
 
 @Component
-public class ControllerDrawMapper {
-    public ServiceDrawSaveDto convertToServiceDrawSaveDto(DrawSaveRequest drawSaveRequest) {
+public class DrawMapper {
+    public ServiceDrawSaveDto toDrawSaveDto(DrawSaveRequest drawSaveRequest) {
         return new ServiceDrawSaveDto(
                 drawSaveRequest.sneakerId(),
                 drawSaveRequest.startDate(),
@@ -16,5 +16,4 @@ public class ControllerDrawMapper {
                 drawSaveRequest.sneakerItems()
         );
     }
-
 }
